Removed unused base classes
